requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-debug-toolbar==3.2.1
django-elasticsearch-dsl==7.2.0
django-extensions==3.1.1
django-filter==1.1.0
django-haystack==2.8.1
django-material==1.9.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
elasticsearch==7.13.4
elasticsearch-dsl==7.4.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
material==0.1
openapi-codec==1.3.2
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
Whoosh==2.7.4
wincertstore==0.2
